Moved utility functions to tripper.utils
