+++ docs/api_reference/utils.md
@@ -1,0 +1,3 @@
+# utils
+
+::: tripper.utils
