+++ tripper/__init__.py
@@ -3,8 +3,7 @@ pattern.
 
 See the README.md file for a description for how to use this package.
 """
-from .errors import NamespaceError, NoSuchIRIError, TriplestoreError, UniquenessError
-from .literal import Literal, en, parse_literal
+from .literal import Literal
 from .namespace import (
     DC,
     DCTERMS,
@@ -22,19 +21,12 @@ from .namespace import (
     XSD,
     Namespace,
 )
-from .triplestore import Triplestore, split_iri
+from .triplestore import Triplestore
 
 __version__ = "0.1.1"
 
 __all__ = (
-    "NamespaceError",
-    "NoSuchIRIError",
-    "TriplestoreError",
-    "UniquenessError",
-    #
     "Literal",
-    "en",
-    "parse_literal",
     #
     "DC",
     "DCTERMS",
@@ -53,5 +45,4 @@ __all__ = (
     "Namespace",
     #
     "Triplestore",
-    "split_iri",
 )
